refactor(#3263): extract TextEncoder/TextDecoder helpers into text-encoding-native.ts (#3182)
Behaviour-preserving god-file split. Moves ensureTextEncodingHelpers and its
private helper ensureEncodeIntoResultStruct verbatim (~642 LOC) out of the
7,461-LOC native-strings.ts into a new sibling module
src/codegen/text-encoding-native.ts. Pure move, no logic changes.

- native-strings.ts: 7,461 -> 6,811 LOC; imports nothing back (no cycle).
- text-encoding-native.ts imports ensureNativeStringHelpers from native-strings.js
  plus type/registry/func-space deps (paths unchanged, sibling module).
- expressions/calls.ts re-pointed its single ensureTextEncodingHelpers import.

Proof:
- npx tsc --noEmit -> 0 errors.
- prove-emit-identity check -> IDENTICAL (39/39 gc/standalone/wasi).
- All relocation-shift ratchets pass with zero allowances (move fully conserved).
- tests/issue-3263.test.ts smoke test green (standalone + wasi).

Subtask of #3182.

✓
